requirements.txt
cudatoolkit==11.3.1
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
pillow==9.4.0
pip==22.3.1
python==3.7.16
pytorch==1.10.0
requests==2.31.0
scikit-learn==1.0.2
sqlite==3.40.1
tqdm==4.64.1
urllib3==1.26.14
